build(dashboardlog): restructure auditlogs folder in shared
